+++ src/Ext/TechnoType/Body.cpp
@@ -14,6 +14,14 @@
 TechnoTypeExt::ExtContainer TechnoTypeExt::ExtMap;
 bool TechnoTypeExt::SelectWeaponMutex = false;
 
+void TechnoTypeExt::ExtData::Initialize()
+{
+	auto pThis = this->OwnerObject();
+
+	if (pThis->WhatAmI() == AircraftTypeClass::AbsID)
+		this->CustomMissileTakeoffAnim = AnimTypeClass::Find("V3TAKOFF");
+}
+
 void TechnoTypeExt::ExtData::ApplyTurretOffset(Matrix3D* mtx, double factor)
 {
 	// Does not verify if the offset actually has all values parsed as it makes no difference, it will be 0 for the unparsed ones either way.
@@ -1204,6 +1212,10 @@ void TechnoTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 
 	this->LandingAnim.Read(exINI, pSection, "LandingAnim");
 
+	this->IsCruiseMissile.Read(exINI, pSection, "Missile.Cruise");
+	this->CustomMissileTakeoffAnim.Read(exINI, pSection, "Missile.TakeOffAnim");
+	this->CustomMissileTakeoffSeparation.Read(exINI, pSection, "Missile.TakeOffSeparation");
+
 	// Ares 0.2
 	this->RadarJamRadius.Read(exINI, pSection, "RadarJamRadius");
 
@@ -1951,6 +1963,10 @@ void TechnoTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->FlyNoWobbles)
 
 		.Process(this->LandingAnim)
+
+		.Process(this->IsCruiseMissile)
+		.Process(this->CustomMissileTakeoffAnim)
+		.Process(this->CustomMissileTakeoffSeparation)
 		;
 }
 void TechnoTypeExt::ExtData::LoadFromStream(PhobosStreamReader& Stm)